ic 705
This is to allow multiple programs access to the radio at the same time.  please be advised that the byid used is not the same for every radio and will need to be changed for your radio. this script can be used for any radio by changing the rigcrtl number and byid also make sure that the baud rate is correct.